xyz point clouds
Various reconstructions using the StomoVer3 code.  The x, y and z coordinates are in units of pixels, where x and y have the same sampling as the tilt series if images input into StomoVer3.

The format is a contiguous stream of 32-bit binary floating point numbers in the form x1, y1, z1, x2, y2, z2, with no delimiters, where the numbers 1, 2 refer to the first point, point number 2 etc.